Extract AbstractOTLPTransportAction
Extract common transport action logic into AbstractOTLPTransportAction base class,
including bulk request handling, success/partial-success/failure response building,
and the ProcessingContext interface. Introduce shared OTLPActionRequest/OTLPActionResponse
types replacing the inner MetricsRequest/MetricsResponse classes.
